# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DrépaConnect API — plateforme de lutte contre la drépanocytose
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients (doctor view)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: sickleType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create patient profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Patient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"

  # CRISES
...
components:
  schemas:
...
    Patient:
      type: object
      required: [id, userId, sickleType, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        sickleType:
          type: string
          enum: [SS, AS, SC, AA, AC, CC]
        bloodType:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        allergies:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PatientInput:
      type: object
      required: [sickleType]
      properties:
        sickleType:
          type: string
          enum: [SS, AS, SC, AA, AC, CC]
        dateOfBirth:
          type: ["string", "null"]
        bloodType:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        allergies:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    PatientUpdate:
      type: object
      properties:
        sickleType:
          type: string
          enum: [SS, AS, SC, AA, AC, CC]
        dateOfBirth:
          type: ["string", "null"]
        bloodType:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        allergies:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # CRISIS
...